Tarefa1_2022li1g102

Description

Módulo para a realização da Tarefa 1 do projeto de LI1 em 2022/23.

Documentation

validaObstaculo :: Mapa -> Bool

A funçao validaObstaculo é uma função auxiliar da função principal "mapaVálido" que recebe um mapa e verifica se os obstáculos nele contidos estão de acordo com o tipo de terreno, returnando True caso os obstaculos correspondam ao terreno fornecido

>>> validaObstaculo (Mapa 3 [(Estrada 1, [Nenhum, Carro, Tronco])])
False

>>> validaObstaculo (Mapa 3 [(Estrada 1, [Nenhum, Carro, Nenhum])])
True

validaRios :: Mapa -> Bool

A funçao validaRios é uma função auxiliar da função mapaVálidos e irá verificar se rios contiguos têm velocidades de sinal simétrico, indicando que os rios têm sentidos opostos. Caso esta condição se verifique, é returnado True

>>> validaRios (Mapa 3 [(Rio 1, [Nenhum, Carro, Tronco]),(Rio 1, [Nenhum,Nenhum, Nenhum])])
False

>>> validaRios (Mapa 3 [(Rio 1, [Nenhum, Carro, Tronco]),(Rio (-1), [Nenhum,Nenhum, Nenhum])])
True

validaComprimentos :: Mapa -> Bool

A função validaComprimentos irá verificar se Carros têm no máximo 3 unidades de comprimento e se Troncos têm no máximo 5. Es caso afirmatico, retorna True

>>> validaComprimentos (Mapa 4 [(Estrada 1, [Carro, Carro, Carro, Nenhum ]),(Rio (-1), [Nenhum,Nenhum, Nenhum])])
False

>>> validaComprimentos (Mapa 4 [(Estrada 1, [Carro, Carro, Nenhum, Nenhum ]),(Rio (-1), [Nenhum,Nenhum, Nenhum])])
True

validaComprimentosAux1 :: Mapa -> Bool

Função Auxiliar da função validaComprimentos

validaComprimentosAux2 :: Mapa -> Bool

Função Auxiliar da função validaComprimentos

validaLinha :: Mapa -> Bool

A função validaLinha verifica se, em cada lista de obstaculos, existe pelo menos um obstáculo Nenhum, para que seja possivel ao jogador avançar no mapa. Em caso afirmatico retorna True

>>> validaLinha (Mapa 4 [(Estrada 1, [Carro, Carro, Carro, Carro ]),(Rio (-1), [Nenhum,Nenhum, Nenhum])])
False

>>> validaLinha (Mapa 4 [(Estrada 1, [Carro, Carro, Nenhum, Nenhum]),(Rio (-1), [Nenhum,Nenhum, Nenhum])])
True

validaLargura :: Mapa -> Bool

A função validaLargura verifica se a largura fornecida corresponde ao comprimento da lista de obstáculos. Caso se confirme, retorna True

>>> validaLargura (Mapa 3 [(Estrada 1, [Carro, Carro, Nenhum, Nenhum]),(Rio (-1), [Nenhum,Nenhum, Nenhum])])
False

>>> validaLargura (Mapa 4 [(Estrada 1, [Carro, Carro, Nenhum, Nenhum]),(Rio (-1), [Nenhum,Nenhum, Nenhum])])
True

validaMapa :: Mapa -> Bool

A função validaMapa verifica se não existem mais do que 5 estradas/ relvas consecutivas e se não existem mais do que 4 rios consecutivos

>>> validaMapa (Mapa 4 [(Estrada 1, [Carro, Carro, Nenhum, Nenhum]),(Estrada 1, [Nenhum,Nenhum, Nenhum, Nenhum]),(Estrada 1, [Nenhum,Nenhum, Nenhum, Nenhum]),(Estrada 1, [Nenhum,Nenhum, Nenhum, Nenhum]),(Estrada 1, [Nenhum,Nenhum, Nenhum, Nenhum]),(Estrada 1, [Nenhum,Nenhum, Nenhum, Nenhum])])
False

>>> validaMapa (Mapa 4 [(Estrada 1, [Carro, Carro, Nenhum, Nenhum]),(Estrada 1, [Nenhum,Nenhum, Nenhum, Nenhum])])
True

mapaVálido :: Mapa -> Bool

A função principal mapaVálido recebe um Mapa e, servindo-se das funções auxiliares acima, retorna o valor da interseção dos seus valores lógicos, validando assim o mapa recebido.