Tarefa3_2022li1g102

Description

Módulo para a realização da Tarefa 3 do projeto de LI1 em 2022/23.

Documentation

indices :: Eq a => Int -> a -> [a] -> [Int]

Função auxiliar da função elemIndices' que retorna a posiçao do elemento desejado na lista recebida

elemIndices' :: Eq a => a -> [a] -> [Int]

A função elemIndices' retorna a lista com o numero da posição onde o elemento x se encontra na lista l

aux :: Velocidade -> [Obstaculo] -> Int -> [Obstaculo]

Função auxiliar da função moveObs 1ue, a partir de uma velocidade, reordena a lista de obstaculos

moveObs :: [(Terreno, [Obstaculo])] -> [(Terreno, [Obstaculo])]

Função auxiliar da função animaJogo, que movimenta os obstaculos de uma linha do mapa.

moveJogador :: Jogo -> Jogada -> Jogo

A função moveJogador altera as coordenadas do jogador, consoante a jogada recebida e , ao mesmo tempo, reeordena os obstaculos de uma certa linha.

moveJogadorAux :: Jogo -> Jogada -> Jogo

A função moveJogadorAux move o jogador consoante a jogada fornecida

moveEmTronco :: Jogo -> Jogada -> Jogo

A função moveEmTronco move o Jogador caso este se encontre num rio, consoante os criterios definidos

verificaRio :: Mapa -> Int -> Int

a função verificaLinha verifica se o jogador se encontra num rio

posAtropelamento :: Jogo -> Jogo

A função pósAtropelamento lida com o caso em que o atropelamento é confirmado (valor True na função atropelamento) e move os objetos da lonha consoante os critérios definidos

atropelamento :: Jogo -> Jogada -> Bool

A função atropelamento compara a posição de um Carro com a posição do jogador e retorna o valor True caso estes coincidam. Esta função apenas se aplica a terrenos do tipo Estrada, uma vez que não faz sentido o jogador ser atropelado noutro tipo de terrenos

animaJogo :: Jogo -> Jogada -> Jogo

A função animaJogo movimenta os obstáculos (de acordo com a velocidade do terreno em que se encontram), e o personagem, de acordo com a jogada dada