Tarefa4_2022li1g102

Description

Módulo para a realização da Tarefa 4 do projeto de LI1 em 2022/23.

Documentation

jogoTerminou

Arguments

:: Jogo
Argumento que indica as coordenadas do jogador e o cenário onde se encontra

-> Bool
Resultado

A função jogoTerminou indica se o jogador perdeu o jogo, onde True significa que sim, quando se encontra fora do mapa, na água ou "debaixo" de um carro.

Exemplos de utilização:

>>> jogoTerminou (Jogo (Jogador (0,1)) (Mapa 3 [(Relva, [Nenhum, Arvore, Nenhum])]))
True
>>> jogoTerminou (Jogo (Jogador (1,1)) (Mapa 3 [(Estrada 3,[Nenhum, Carro, Nenhum])]))
False

eRio :: Terreno -> Bool

eEstrada :: Terreno -> Bool